ldr-hdr, low precision for shaders

+++ Resources/Engine/Shaders/defaultFragment.shader
@@ -18,10 +18,10 @@
 	<!--
 		#version 300 es
 		precision highp float;
-		precision highp sampler2D;
-		precision highp samplerCube;
+		precision mediump sampler2D;
+		precision mediump samplerCube;
 		precision highp sampler2DArray;
-		precision highp int;
+		precision lowp int;
 
 		uniform sampler2D s_texture0; // color
 		uniform sampler2D s_texture1; // emissive
