+++ Resources/Engine/Shaders/defaultFragment_alphamask.shader
@@ -7,10 +7,10 @@
 	<!--
 		#version 300 es
 		precision highp float;
-		precision highp sampler2D;
-		precision highp samplerCube;
+		precision mediump sampler2D;
+		precision mediump samplerCube;
 		precision highp sampler2DArray;
-		precision highp int;
+		precision lowp int;
 	-->
 	</source>
 </shader>
